science-scraper/requirements.txt
requests
beautifulsoup4
biopython
elsapy
pandas
python-dotenv
